requirements.txt
textual
python-dateutil
sqlalchemy
python-dotenv
pymysql
